+++ src/steps/resource-manager/network/index.test.ts
@@ -19,6 +19,12 @@ import { createMockAzureStepExecutionContext } from '../../../../test/createMock
 import { MonitorEntities } from '../monitor/constants';
 import { IntegrationConfig } from '../../../types';
 import { ACCOUNT_ENTITY_TYPE } from '../../active-directory';
+import {
+  LOAD_BALANCER_ENTITY_CLASS,
+  LOAD_BALANCER_ENTITY_TYPE,
+  PUBLIC_IP_ADDRESS_ENTITY_CLASS,
+  PUBLIC_IP_ADDRESS_ENTITY_TYPE,
+} from './constants';
 
 const GUID_REGEX = new RegExp(
   '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$',
@@ -97,21 +103,26 @@ let recording: Recording;
 let instanceConfig: IntegrationConfig;
 
 afterEach(async () => {
-  await recording.stop();
+  if (recording) {
+    await recording.stop();
+  }
 });
 
 test('network steps', async () => {
   instanceConfig = {
     clientId: process.env.CLIENT_ID || 'clientId',
     clientSecret: process.env.CLIENT_SECRET || 'clientSecret',
-    directoryId: 'bcd90474-9b62-4040-9d7b-8af257b1427d',
-    subscriptionId: '40474ebe-55a2-4071-8fa8-b610acdd8e56',
+    directoryId: '4a17becb-fb42-4633-b5c8-5ab66f28d195',
+    subscriptionId: '87f62f44-9dad-4284-a08f-f2fb3d8b528a',
     developerId: 'keionned',
   };
 
   recording = setupAzureRecording({
     directory: __dirname,
     name: 'network-steps',
+    options: {
+      recordFailedRequests: true,
+    },
   });
 
   const resouceGroupEntity: Entity = {
@@ -141,6 +152,8 @@ test('network steps', async () => {
   await fetchVirtualNetworks(context);
   await fetchLoadBalancers(context);
   await buildSecurityGroupRuleRelationships(context);
+
+  // console.log(JSON.stringify(context.jobState.collectedEntities, null, 2))
 
   expect(context.jobState.collectedEntities).toContainOnlyGraphObjects(
     {
@@ -174,7 +187,7 @@ test('network steps', async () => {
       region: 'eastus',
       resourceGroup: 'j1dev',
       'tag.environment': 'j1dev',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev`,
     },
     {
       CIDR: '10.0.2.0/24',
@@ -191,7 +204,7 @@ test('network steps', async () => {
       public: false,
       region: 'eastus',
       resourceGroup: 'j1dev',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev/subnets/j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev/subnets/j1dev`,
     },
     {
       CIDR: '10.0.3.0/24',
@@ -208,7 +221,7 @@ test('network steps', async () => {
       public: false,
       region: 'eastus',
       resourceGroup: 'j1dev',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev/subnets/j1dev_priv_one`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev/subnets/j1dev_priv_one`,
     },
     {
       _class: 'IpAddress',
@@ -227,7 +240,7 @@ test('network steps', async () => {
       sku: 'Basic',
       'tag.environment': 'j1dev',
       type: 'Microsoft.Network/publicIPAddresses',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev`,
     },
     {
       _class: 'NetworkInterface',
@@ -251,7 +264,7 @@ test('network steps', async () => {
       // Only defined when bound to a VM
       // virtualMachineId:
       //   `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Compute/virtualMachines/j1dev`,
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/networkInterfaces/j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/networkInterfaces/j1dev`,
     },
     {
       _class: ['Firewall'],
@@ -268,7 +281,43 @@ test('network steps', async () => {
       region: 'eastus',
       resourceGroup: 'j1dev',
       'tag.environment': 'j1dev',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/networkSecurityGroups/j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/networkSecurityGroups/j1dev`,
+    },
+    {
+      id: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev_lb_ip`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev_lb_ip`,
+      _class: PUBLIC_IP_ADDRESS_ENTITY_CLASS,
+      _type: PUBLIC_IP_ADDRESS_ENTITY_TYPE,
+      _rawData: [expect.objectContaining({ name: 'default' })],
+      displayName: 'j1dev_lb_ip',
+      public: true,
+      // Dynamic address may be undefined when the address is not bound to anything
+      // publicIp: '52.188.119.30',
+      // publicIpAddress: '52.188.119.30',
+      region: 'eastus',
+      resourceGroup: 'j1dev',
+      resourceGuid: expect.stringMatching(GUID_REGEX),
+      sku: 'Basic',
+      'tag.environment': 'j1dev',
+      type: 'Microsoft.Network/publicIPAddresses',
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev_lb_ip`,
+    },
+    {
+      id: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer`,
+      _class: [LOAD_BALANCER_ENTITY_CLASS],
+      _type: LOAD_BALANCER_ENTITY_TYPE,
+      category: ['network'],
+      function: ['load-balancing'],
+      resourceGuid: expect.stringMatching(GUID_REGEX),
+      resourceGroup: 'j1dev',
+      displayName: 'TestLoadBalancer',
+      type: 'Microsoft.Network/loadBalancers',
+      region: 'eastus',
+      publicIp: expect.any(Array),
+      privateIp: expect.any(Array),
+      public: true,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer`,
     },
 
     // VPC default, westus
@@ -288,7 +337,7 @@ test('network steps', async () => {
       region: 'westus',
       resourceGroup: 'j1dev',
       'tag.environment': 'j1dev',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev_two`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev_two`,
     },
     {
       CIDR: '10.0.3.0/24',
@@ -305,7 +354,7 @@ test('network steps', async () => {
       public: false,
       region: 'westus',
       resourceGroup: 'j1dev',
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev_two/subnets/j1dev_priv_two`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev_two/subnets/j1dev_priv_two`,
     },
     // Diagnostic Settings
     {
@@ -325,7 +374,7 @@ test('network steps', async () => {
       'retentionPolicy.enabled': true,
       serviceBusRuleId: null,
       storageAccountId: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set`,
       workspaceId: null,
     },
     {
@@ -345,7 +394,67 @@ test('network steps', async () => {
       'retentionPolicy.enabled': true,
       serviceBusRuleId: null,
       storageAccountId: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
-      webLink: `https://portal.azure.com/#@kanddventurepartnersgmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set`,
+      workspaceId: null,
+    },
+    {
+      _class: MonitorEntities.DIAGNOSTIC_LOG_SETTING._class,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set/logs/DDoSMitigationFlowLogs/true/1/true`,
+      _type: MonitorEntities.DIAGNOSTIC_LOG_SETTING._type,
+      _rawData: [expect.objectContaining({ name: 'default' })],
+      category: 'DDoSMitigationFlowLogs',
+      displayName: 'j1dev_pub_ip_diag_set',
+      enabled: true,
+      eventHubAuthorizationRuleId: null,
+      eventHubName: null,
+      id: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set/logs/DDoSMitigationFlowLogs/true/1/true`,
+      logAnalyticsDestinationType: null,
+      name: 'j1dev_pub_ip_diag_set',
+      'retentionPolicy.days': 1,
+      'retentionPolicy.enabled': true,
+      serviceBusRuleId: null,
+      storageAccountId: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set`,
+      workspaceId: null,
+    },
+    {
+      id: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set/logs/LoadBalancerAlertEvent/true/1/true`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set/logs/LoadBalancerAlertEvent/true/1/true`,
+      _class: MonitorEntities.DIAGNOSTIC_LOG_SETTING._class,
+      _type: MonitorEntities.DIAGNOSTIC_LOG_SETTING._type,
+      _rawData: [expect.objectContaining({ name: 'default' })],
+      category: 'LoadBalancerAlertEvent',
+      displayName: 'j1dev_lb_diag_set',
+      enabled: true,
+      eventHubAuthorizationRuleId: null,
+      eventHubName: null,
+      logAnalyticsDestinationType: null,
+      name: 'j1dev_lb_diag_set',
+      'retentionPolicy.days': 1,
+      'retentionPolicy.enabled': true,
+      serviceBusRuleId: null,
+      storageAccountId: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set`,
+      workspaceId: null,
+    },
+    {
+      id: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set/logs/VMProtectionAlerts/true/1/true`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set/logs/VMProtectionAlerts/true/1/true`,
+      _class: MonitorEntities.DIAGNOSTIC_LOG_SETTING._class,
+      _type: MonitorEntities.DIAGNOSTIC_LOG_SETTING._type,
+      _rawData: [expect.objectContaining({ name: 'default' })],
+      category: 'VMProtectionAlerts',
+      displayName: 'j1dev_vn_diag_set',
+      enabled: true,
+      eventHubAuthorizationRuleId: null,
+      eventHubName: null,
+      logAnalyticsDestinationType: null,
+      name: 'j1dev_vn_diag_set',
+      'retentionPolicy.days': 1,
+      'retentionPolicy.enabled': true,
+      serviceBusRuleId: null,
+      storageAccountId: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      webLink: `https://portal.azure.com/#@knnderoussellegmail.onmicrosoft.com/resource/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set`,
       workspaceId: null,
     },
   );
@@ -791,6 +900,22 @@ test('network steps', async () => {
       _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev_two`,
       displayName: 'HAS',
     },
+    {
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev|has|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer`,
+      _type: 'azure_resource_group_has_lb',
+      _class: 'HAS',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer`,
+      displayName: 'HAS',
+    },
+    {
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev|has|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev_lb_ip`,
+      _type: 'azure_resource_group_has_public_ip',
+      _class: 'HAS',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev_lb_ip`,
+      displayName: 'HAS',
+    },
     // Diagnostic Settings
     {
       _class: 'HAS',
@@ -809,6 +934,30 @@ test('network steps', async () => {
       displayName: 'HAS',
     },
     {
+      _class: 'HAS',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/publicIPAddresses/j1dev|has|/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set/logs/DDoSMitigationFlowLogs/true/1/true`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set/logs/DDoSMitigationFlowLogs/true/1/true`,
+      _type: 'azure_resource_has_diagnostic_log_setting',
+      displayName: 'HAS',
+    },
+    {
+      _class: 'HAS',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/loadBalancers/TestLoadBalancer|has|/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set/logs/LoadBalancerAlertEvent/true/1/true`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set/logs/LoadBalancerAlertEvent/true/1/true`,
+      _type: 'azure_resource_has_diagnostic_log_setting',
+      displayName: 'HAS',
+    },
+    {
+      _class: 'HAS',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Network/virtualNetworks/j1dev|has|/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set/logs/VMProtectionAlerts/true/1/true`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set/logs/VMProtectionAlerts/true/1/true`,
+      _type: 'azure_resource_has_diagnostic_log_setting',
+      displayName: 'HAS',
+    },
+    {
       _class: 'USES',
       _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set/logs/NetworkSecurityGroupRuleCounter/true/1/true`,
       _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set/logs/NetworkSecurityGroupRuleCounter/true/1/true|uses|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
@@ -820,6 +969,30 @@ test('network steps', async () => {
       _class: 'USES',
       _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set/logs/NetworkSecurityGroupEvent/true/1/true`,
       _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/networksecuritygroups/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_net_sec_grp_set/logs/NetworkSecurityGroupEvent/true/1/true|uses|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      _type: 'azure_diagnostic_log_setting_uses_storage_account',
+      displayName: 'USES',
+    },
+    {
+      _class: 'USES',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set/logs/DDoSMitigationFlowLogs/true/1/true`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/publicipaddresses/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_pub_ip_diag_set/logs/DDoSMitigationFlowLogs/true/1/true|uses|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      _type: 'azure_diagnostic_log_setting_uses_storage_account',
+      displayName: 'USES',
+    },
+    {
+      _class: 'USES',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set/logs/LoadBalancerAlertEvent/true/1/true`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/loadbalancers/testloadbalancer/providers/microsoft.insights/diagnosticSettings/j1dev_lb_diag_set/logs/LoadBalancerAlertEvent/true/1/true|uses|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
+      _type: 'azure_diagnostic_log_setting_uses_storage_account',
+      displayName: 'USES',
+    },
+    {
+      _class: 'USES',
+      _fromEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set/logs/VMProtectionAlerts/true/1/true`,
+      _key: `/subscriptions/${instanceConfig.subscriptionId}/resourcegroups/j1dev/providers/microsoft.network/virtualnetworks/j1dev/providers/microsoft.insights/diagnosticSettings/j1dev_vn_diag_set/logs/VMProtectionAlerts/true/1/true|uses|/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
       _toEntityKey: `/subscriptions/${instanceConfig.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/${instanceConfig.developerId}j1dev`,
       _type: 'azure_diagnostic_log_setting_uses_storage_account',
       displayName: 'USES',